portal26_ngrok_agent/requirements.txt
google-adk
litellm>=1.75.5
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-grpc
opentelemetry-instrumentation
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-google-genai
python-dotenv
